Improve Java installation and locating in install.sh
Let the user decide if he wants to install Java or not, and ask him for his distro if he does (currently only Debian/Ubuntu),
Check if the JAVA_HOME is correct and allow the user to enter it manually.
Also changed the "echo -n ...; read ..." to "echo -e -p ... ..." - this enables path completion and arrow keys.